requirements.txt
black==23.9.1
Django==4.2.5
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
flake8==6.1.0
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
